Added swagger to AWS VM Service
