+++ aws-vm-service/requirements.txt
@@ -1,3 +1,5 @@
 Flask == 1.0.2
 boto3 == 1.11.9
 requests == 2.21.0
+connexion == 2.6.0
+swagger-ui-bundle==0.0.2
